requirements.txt
google-auth
requests
fastapi
uvicorn